# Conflicts:
#	gradle.properties
#	src/generated/resources/.cache/103d9f3f36b01595f1aa5172191e60eff02e6924
#	src/generated/resources/.cache/16e3109ff91cafd392f76e76b7877bb2ffcc596a
#	src/generated/resources/.cache/59eb3dbb5f86130e09b3c62d89b9525ee01cf52d
#	src/generated/resources/.cache/753083c6ae434088f8bc892fc69c207b792eed97
#	src/generated/resources/.cache/81ff2ae0d1283b52cda2124a9e738f4aead0b23d
#	src/generated/resources/.cache/dbe3d85901421efbc772135cedc481ee976ed44b
#	src/generated/resources/.cache/fe9ab2fe767116f93410e684d2796197decdd205
#	src/generated/resources/data/minecraft/tags/block/mineable/pickaxe.json
#	src/main/java/quek/undergarden/data/UGBiomeTags.java
#	src/main/java/quek/undergarden/data/UGBlockTags.java
#	src/main/java/quek/undergarden/data/UGItemModels.java
#	src/main/java/quek/undergarden/event/UndergardenClientEvents.java
#	src/main/java/quek/undergarden/registry/UGBiomes.java
#	src/main/java/quek/undergarden/registry/UGBlocks.java
#	src/main/java/quek/undergarden/registry/UGParticleTypes.java
#	src/main/java/quek/undergarden/registry/UGSoundTypes.java
#	src/main/java/quek/undergarden/registry/UGTags.java
#	src/main/resources/META-INF/accesstransformer.cfg

+++ src/main/java/quek/undergarden/network/UthericInfectionPacket.java
@@ -11,18 +11,18 @@ import net.neoforged.neoforge.network.handling.IPayloadContext;
 import quek.undergarden.Undergarden;
 import quek.undergarden.registry.UGAttachments;
 
-public record UthericInfectionPacket(int entityID, int infectionLevel) implements CustomPacketPayload {
+public record UthericInfectionPacket(int entityID, double infectionLevel) implements CustomPacketPayload {
 
 	public static final Type<UthericInfectionPacket> TYPE = new Type<>(ResourceLocation.fromNamespaceAndPath(Undergarden.MODID, "utheric_infection"));
 	public static final StreamCodec<FriendlyByteBuf, UthericInfectionPacket> STREAM_CODEC = CustomPacketPayload.codec(UthericInfectionPacket::write, UthericInfectionPacket::new);
 
 	public UthericInfectionPacket(FriendlyByteBuf buf) {
-		this(buf.readInt(), buf.readInt());
+		this(buf.readInt(), buf.readDouble());
 	}
 
 	public void write(FriendlyByteBuf buffer) {
 		buffer.writeInt(this.entityID());
-		buffer.writeInt(this.infectionLevel());
+		buffer.writeDouble(this.infectionLevel());
 	}
 
 	@Override
